+++ src/index.js
@@ -29,6 +29,7 @@ import Row from "./Row";
 import Select from "./Select";
 import SeparatorList from "./SeparatorList";
 import Sidebar from "./Sidebar";
+import Slider from "./Slider";
 import Tabs from "./Tabs";
 import Tag from "./Tag";
 import TextInput from "./TextInput";
@@ -72,6 +73,7 @@ export {
   Select,
   SeparatorList,
   Sidebar,
+  Slider,
   Tabs,
   Tag,
   TextInput,