+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.33.0](https://github.com/narmi/design_system/compare/v3.32.6...v3.33.0) (2024-05-01)
+
+
+### Features
+
+* **Slider:** Add slider component ([21d276d](https://github.com/narmi/design_system/commit/21d276da7abfb41f744374152e5b7a6ebfaac24c))
+
+### [3.32.6](https://github.com/narmi/design_system/compare/v3.32.5...v3.32.6) (2024-05-01)
+
+
+### Bug Fixes
+
+* **IconButton:** use default browser outline styling on focus ([841e1eb](https://github.com/narmi/design_system/commit/841e1eb70b714fc4fa3fde8654d550644df9bece))
+* **IconButton:** use focus-visible instead ([3e8aa40](https://github.com/narmi/design_system/commit/3e8aa40450ec797387dfec7b726ff90e029dc86e))
+
 ### [3.32.5](https://github.com/narmi/design_system/compare/v3.32.4...v3.32.5) (2024-04-25)